Refactor to separate app from main
